+++ test/testRoot/src/app.js
@@ -12,3 +12,5 @@ comment
 comabc
 
 ccc
+
+ts