+++ test/testRoot/src/package/test.ts
@@ -1,0 +1,8 @@
+export function tssss() {
+
+}
+
+
+export function ts() {
+
+}