+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to the "vscode-js-import" extension will be documented in th
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.3.2] - 2017-05-26
+- support ts file in source code
+- fix option filesToScan inactive bug
+- add error msg when attempt to import mutiple default variable
+
 ## [0.3.1] - 2017-05-26
 - support as in import statement (currently not support line feed)
 - clean up import although is already import